pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "termchan"
version = "0.1.0"
description = "browse the boards from your terminal"
requires-python = ">=3.11"

dependencies = [
    "textual>=0.50.0",
    "httpx>=0.27.0",
    "textual-image>=0.6.0",
    "Pillow>=10.0.0",
]

[project.scripts]
termchan = "termchan.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
